ATO-2464: Combine validate VTR methods into base class
- Both of these methods were doing the same thing with a list of VectorOfTrusts, we just needed to handle the case where we could not extract the vtr,
from the request object (when a ParseException is thrown)
