+++ oidc-api/src/main/java/uk/gov/di/authentication/oidc/validators/BaseAuthorizeValidator.java
@@ -149,6 +149,41 @@ public abstract class BaseAuthorizeValidator {
         return Optional.empty();
     }
 
+    protected Optional<ErrorObject> validateVtr(List<String> authRequestVtr, ClientRegistry client) {
+        try {
+            var vtrList = VectorOfTrust.parseFromAuthRequestAttribute(authRequestVtr);
+            var levelOfConfidenceValues = VectorOfTrust.getRequestedLevelsOfConfidence(vtrList);
+            if (!client.getClientLoCs().containsAll(levelOfConfidenceValues)) {
+                logErrorInProdElseWarn(
+                        String.format(
+                                "Level of confidence values have been requested which this client is not permitted to request. Level of confidence values in request: %s",
+                                levelOfConfidenceValues));
+                return Optional.of(
+                        new ErrorObject(
+                                OAuth2Error.INVALID_REQUEST_CODE,
+                                "Request vtr is not permitted"));
+            }
+            var vtrError = errorIfIdentityLoCAndIdentityUnsupported(vtrList, client);
+            if (vtrError.isPresent()) {
+                return vtrError;
+            }
+            if (vtrList.get(0).containsLevelOfConfidence()
+                    && !ipvCapacityService.isIPVCapacityAvailable()
+                    && !client.isTestClient()) {
+                return Optional.of(OAuth2Error.TEMPORARILY_UNAVAILABLE);
+            }
+        } catch (IllegalArgumentException e) {
+            logErrorInProdElseWarn(
+                    String.format(
+                            "vtr in AuthRequest is not valid. vtr in request: %s. IllegalArgumentException: %s",
+                            authRequestVtr, e));
+            return Optional.of(
+                    new ErrorObject(
+                            OAuth2Error.INVALID_REQUEST_CODE, "Request vtr not valid"));
+        }
+        return Optional.empty();
+    }
+
     protected Optional<ErrorObject> errorIfIdentityLoCAndIdentityUnsupported(
             List<VectorOfTrust> vtrList, ClientRegistry client) {
         List<LevelOfConfidence> identityLoCs =
